requirements.txt
nltk==3.8.1
PyPDF2==3.0.1
flask==3.0.3
python-dotenv==1.0.1
openai>=1.30.0
